requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
colorama==0.4.6
constantly==15.1.0
coverage==7.2.7
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.2
flake8==6.1.0
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
mccabe==0.7.0
msgpack==1.0.5
multidict==6.0.4
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
redis==4.6.0
selenium==4.11.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
zope.interface==6.0
